TaskEvent, UiEvent, Routes, TaskItem, and SharedViewModel Changes
I added an event handler for Task Events and UI events that can handle things like deleting tasks, undo delete, editing tasks, etc. I added a Routes object for screen navigation. I created a TaskItem file for a composable so that Anita can start playing with layout of editing the tasks(It should be able to just get called in the dashboard instead of the other file). In the SharedViewModel I added the event handling items for on click events and whatnot.