Added TaskScreen and fixed Dependency
I added a new TaskScreen since the TaskItem wasn't enough to be viewed. This still isn't completely implemented, I still need to add an add_edit_task screen so that this can all work together. There was a hilt dependency that was missing that is now added so the hilt view model should work now too.