Project Builds / App is Crashing
I added the screens and all the implementations for the tasks, I fixed an issue with not having an application class set up for daggerhilt to recognize. I also added an app module class for dependency injection. The project builds but the app is crashing right now. In logcat it talks about an issue with dagger hilt not recognizing the application still so I am still working on it. *Project Builds* *App Crashes*